scripts/requirements2.txt
streamlit
pandas
numpy
plotly
tensorflow-cpu
scikit-learn
joblib